+++ packages/worker-log/tsup.config.ts
@@ -11,6 +11,6 @@ export default defineConfig({
   sourcemap: true,
   clean: true,
   banner: {
-    js: "/*! libreconsent v1.0.0 | MIT License | SPDX-License-Identifier: MIT */",
+    js: "/*! libreconsent v1.1.0 | MIT License | SPDX-License-Identifier: MIT */",
   },
 });